archival-pipeline/requirements.txt
psycopg2-binary
pandas
python-dotenv
pytest
boto3